requirements.txt
oracledb
pymongo
python-dotenv
Jinja2
python-dateutil